projects/llm_backend/pyproject.toml
[project]
name = "llm-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "langchain-community>=0.3.13",
    "langchain>=0.3.13",
    "openai>=1.58.1",
    "rich>=13.9.4",
    "pyinstaller>=6.11.1",
    "fuzzywuzzy[speedup]>=0.18.0",
    "langchain-openai>=0.2.14",
    "pymupdf4llm>=0.0.17",
    "azure-ai-inference>=1.0.0b9",
    "motor>=3.3.0",
    "pymongo>=4.6.0",
    "pydantic[email]>=2.0.0",
    "pydantic-settings>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "beanie>=1.23.0",
]
